requirements.txt
#ep8>=1.5.6
pylint>=1.2.1
flask==0.10.1
Flask-Assets
#
#flask-bcrypt ===> sudo apt-get install python3.5-dev    easy_install flask-bcrypt
flask-login==0.2.11
flask-mail==0.9.0
flask-migrate
flask-script==2.0.5
flask-wtf==0.9.5
alembic
pytest==2.5.2
webtest==2.0.15
cssmin==0.2.0
gunicorn==0.17.2
sqlalchemy==0.9.3
WTForms-Components
simplejson
xmltodict
Werkzeug
paho-mqtt
django
pymysql
python-socketio
flask-socketio==0.6.0
scss
flask_bcrypt
decorator
infinity
#blockspring
#psycopg2==2.5.2
# mysql -- sudo pip install mysql-connector-python --allow-external mysql-connector-python
#pip install --user --allow-external mysql-connector-python mysql-connector-python


